Huge update overall
Updated most of the methods, separated and assigned each task to it's own method which improved maintainability and readability so much.
Created bunch of helper methods for each functionality, Previously the updateOrder method was very huge now it's helper methods handles most of the work. and changed the price type from float to BigDecimal to reduce errors.